+++ src/PerimeterOfShapes.java
@@ -14,7 +14,7 @@ public class PerimeterOfShapes {
         switch (ch){
             case 1:
                 System.out.println("Enter radius : ");
-                int r = sc.nextInt();
+                double r = sc.nextInt();
                 double pi = 3.14;
                 double perimeter = 2 * pi * r;
                 System.out.println("Perimeter of Circle : " + perimeter);
